move TestBit to translatebits.cc/.h, OutputFormat to wmbus.h
