+++ src/driver_dynamic.cc
@@ -20,6 +20,7 @@
 #include"crc16.h"
 #include"log.h"
 #include"meters_common_implementation.h"
+#include"translatebits.h"
 #include"xmq.h"
 
 #include<assert.h>
