+++ src/translatebits.h
@@ -18,11 +18,20 @@
 #ifndef TRANSLATEBITS_H
 #define TRANSLATEBITS_H
 
+#include"util.h"
+
 #include<cstdint>
 #include<string>
 #include<vector>
 
-#include"util.h"
+enum class TestBit
+{
+    Unknown,
+    Set,
+    NotSet
+};
+
+TestBit toTestBit(const char *s);
 
 struct TriggerBits
 {
